requirements.txt
psycopg2-binary
python-dotenv
pandas
pyarrow
psycopg2